requirements.txt
opencv-python
tqdm
torch
torchvision
scikit-learn
matplotlib
pillow
ipykernel